+++ import_export/formats/base_formats.py
@@ -1,6 +1,7 @@
 from __future__ import unicode_literals
 from django.utils.six import moves
 
+import sys
 import warnings
 import tablib
 
@@ -18,6 +19,21 @@ except ImportError:
         "import support for 'xls' format and xlrd module is not found."
         warnings.warn(xls_warning, ImportWarning)
         XLS_IMPORT = False
+
+
+try:
+    import openpyxl
+    XLSX_IMPORT = True
+except ImportError:
+    try:
+        from tablib.compat import openpyxl
+        XLSX_IMPORT = hasattr(openpyxl, 'load_workbook')
+    except ImportError:
+        xlsx_warning = "Installed `tablib` library does not include"
+        "import support for 'xlsx' format and openpyxl module is not found."
+        warnings.warn(xlsx_warning, ImportWarning)
+        XLSX_IMPORT = False
+
 
 try:
     from importlib import import_module
@@ -37,7 +53,7 @@ class Format(object):
         """
         raise NotImplementedError()
 
-    def export_data(self, dataset):
+    def export_data(self, dataset, **kwargs):
         """
         Returns format representation for given dataset.
         """
@@ -62,7 +78,8 @@ class Format(object):
         return ""
 
     def get_content_type(self):
-        # For content types see http://www.iana.org/assignments/media-types/media-types.xhtml
+        # For content types see
+        # http://www.iana.org/assignments/media-types/media-types.xhtml
         return 'application/octet-stream'
 
     def can_import(self):
@@ -85,17 +102,18 @@ class TablibFormat(Format):
     def get_title(self):
         return self.get_format().title
 
-    def create_dataset(self, in_stream):
+    def create_dataset(self, in_stream, **kwargs):
         data = tablib.Dataset()
-        self.get_format().import_set(data, in_stream)
+        self.get_format().import_set(data, in_stream, **kwargs)
         return data
 
-    def export_data(self, dataset):
-        return self.get_format().export_set(dataset)
+    def export_data(self, dataset, **kwargs):
+        return self.get_format().export_set(dataset, **kwargs)
 
     def get_extension(self):
-        # we support both 'extentions' and 'extensions' because currently tablib's master
-        # branch uses 'extentions' (which is a typo) but it's dev branch already uses 'extension'.
+        # we support both 'extentions' and 'extensions' because currently
+        # tablib's master branch uses 'extentions' (which is a typo) but it's
+        # dev branch already uses 'extension'.
         # TODO - remove this once the typo is fixxed in tablib's master branch
         if hasattr(self.get_format(), 'extentions'):
             return self.get_format().extentions[0]
@@ -119,18 +137,15 @@ class TextFormat(TablibFormat):
         return False
 
 
-class CSV(TablibFormat):
-    """
-    CSV is treated as binary in Python 2.
-    """
+class CSV(TextFormat):
     TABLIB_MODULE = 'tablib.formats._csv'
     CONTENT_TYPE = 'text/csv'
 
-    def get_read_mode(self):
-        return 'rU' if six.PY3 else 'rb'
-
-    def is_binary(self):
-        return False if six.PY3 else True
+    def create_dataset(self, in_stream, **kwargs):
+        if sys.version_info[0] < 3:
+            # python 2.7 csv does not do unicode
+            return super(CSV, self).create_dataset(in_stream.encode('utf-8'), **kwargs)
+        return super(CSV, self).create_dataset(in_stream, **kwargs)
 
 
 class JSON(TextFormat):
@@ -140,7 +155,8 @@ class JSON(TextFormat):
 
 class YAML(TextFormat):
     TABLIB_MODULE = 'tablib.formats._yaml'
-    CONTENT_TYPE = 'text/yaml' # See http://stackoverflow.com/questions/332129/yaml-mime-type
+    # See http://stackoverflow.com/questions/332129/yaml-mime-type
+    CONTENT_TYPE = 'text/yaml'
 
 
 class TSV(TextFormat):
@@ -151,11 +167,6 @@ class TSV(TextFormat):
 class ODS(TextFormat):
     TABLIB_MODULE = 'tablib.formats._ods'
     CONTENT_TYPE = 'application/vnd.oasis.opendocument.spreadsheet'
-
-
-class XLSX(TextFormat):
-    TABLIB_MODULE = 'tablib.formats._xlsx'
-    CONTENT_TYPE = 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'
 
 
 class HTML(TextFormat):
@@ -182,4 +193,29 @@ class XLS(TablibFormat):
         dataset.headers = sheet.row_values(0)
         for i in moves.range(1, sheet.nrows):
             dataset.append(sheet.row_values(i))
+        return dataset
+
+
+class XLSX(TablibFormat):
+    TABLIB_MODULE = 'tablib.formats._xlsx'
+    CONTENT_TYPE = 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'
+
+    def can_import(self):
+        return XLSX_IMPORT
+
+    def create_dataset(self, in_stream):
+        """
+        Create dataset from first sheet.
+        """
+        assert XLSX_IMPORT
+        from io import BytesIO
+        xlsx_book = openpyxl.load_workbook(BytesIO(in_stream))
+
+        dataset = tablib.Dataset()
+        sheet = xlsx_book.active
+
+        dataset.headers = [cell.value for cell in sheet.rows[0]]
+        for i in moves.range(1, len(sheet.rows)):
+            row_values = [cell.value for cell in sheet.rows[i]]
+            dataset.append(row_values)
         return dataset