+++ tests/core/models.py
@@ -1,4 +1,6 @@
 from __future__ import unicode_literals
+import random
+import string
 
 from django.db import models
 from django.utils.encoding import python_2_unicode_compatible
@@ -28,8 +30,9 @@ class Book(models.Model):
     author_email = models.EmailField('Author email', max_length=75, blank=True)
     imported = models.BooleanField(default=False)
     published = models.DateField('Published', blank=True, null=True)
+    published_time = models.TimeField('Time published', blank=True, null=True)
     price = models.DecimalField(max_digits=10, decimal_places=2, null=True,
-            blank=True)
+                                blank=True)
     categories = models.ManyToManyField(Category, blank=True)
 
     def __str__(self):
@@ -43,3 +46,21 @@ class Profile(models.Model):
 
 class Entry(models.Model):
     user = models.ForeignKey('auth.User')
+
+
+class WithDefault(models.Model):
+    name = models.CharField('Default', max_length=75, blank=True,
+                            default='foo_bar')
+
+def random_name():
+    chars = string.ascii_lowercase
+    return ''.join(random.SystemRandom().choice(chars) for _ in range(100))
+
+class WithDynamicDefault(models.Model):
+
+    name = models.CharField('Dyn Default', max_length=100,
+            default=random_name)
+
+
+class WithFloatField(models.Model):
+    f = models.FloatField(blank=True, null=True)