+++ tests/core/tests/widgets_tests.py
@@ -2,9 +2,11 @@
 from __future__ import unicode_literals
 
 from decimal import Decimal
-from datetime import date, datetime
+from datetime import date, datetime, time
 
+from django.test.utils import override_settings
 from django.test import TestCase
+from django.utils import timezone
 
 from import_export import widgets
 
@@ -43,6 +45,11 @@ class DateWidgetTest(TestCase):
     def test_clean(self):
         self.assertEqual(self.widget.clean("13.08.2012"), self.date)
 
+    @override_settings(USE_TZ=True)
+    def test_use_tz(self):
+        self.assertEqual(self.widget.render(self.date), "13.08.2012")
+        self.assertEqual(self.widget.clean("13.08.2012"), self.date)
+
 
 class DateTimeWidgetTest(TestCase):
 
@@ -61,6 +68,15 @@ class DateTimeWidgetTest(TestCase):
         self.assertEqual(self.widget.clean("13.08.2012 18:00:00"),
                          self.datetime)
 
+    @override_settings(USE_TZ=True)
+    def test_use_tz(self):
+        self.assertEqual(self.widget.render(self.datetime),
+                         "13.08.2012 18:00:00")
+        aware_dt = timezone.make_aware(self.datetime,
+                                       timezone.get_default_timezone())
+        self.assertEqual(self.widget.clean("13.08.2012 18:00:00"),
+                         aware_dt)
+
 
 class DateWidgetBefore1900Test(TestCase):
 
@@ -75,6 +91,22 @@ class DateWidgetBefore1900Test(TestCase):
         self.assertEqual(self.widget.clean("13.08.1868"), self.date)
 
 
+class TimeWidgetTest(TestCase):
+
+    def setUp(self):
+        self.time = time(20, 15, 0)
+        self.widget = widgets.TimeWidget('%H:%M:%S')
+
+    def test_render(self):
+        self.assertEqual(self.widget.render(self.time), "20:15:00")
+
+    def test_render_none(self):
+        self.assertEqual(self.widget.render(None), "")
+
+    def test_clean(self):
+        self.assertEqual(self.widget.clean("20:15:00"), self.time)
+
+
 class DecimalWidgetTest(TestCase):
 
     def setUp(self):
@@ -86,6 +118,10 @@ class DecimalWidgetTest(TestCase):
 
     def test_render(self):
         self.assertEqual(self.widget.render(self.value), self.value)
+
+    def test_clean_string_zero(self):
+        self.assertEqual(self.widget.clean("0"), Decimal("0"))
+        self.assertEqual(self.widget.clean("0.0"), Decimal("0"))
 
 
 class IntegerWidgetTest(TestCase):
@@ -99,6 +135,7 @@ class IntegerWidgetTest(TestCase):
 
     def test_clean_string_zero(self):
         self.assertEqual(self.widget.clean("0"), self.value)
+        self.assertEqual(self.widget.clean("0.0"), self.value)
 
 
 class ForeignKeyWidgetTest(TestCase):
@@ -135,9 +172,14 @@ class ManyToManyWidget(TestCase):
         self.assertIn(self.cat1, cleaned_data)
         self.assertIn(self.cat2, cleaned_data)
 
+    def test_clean_typo(self):
+        value = "%s," % self.cat1.pk
+        cleaned_data = self.widget.clean(value)
+        self.assertEqual(len(cleaned_data), 1)
+        self.assertIn(self.cat1, cleaned_data)
+
     def test_render(self):
         self.assertEqual(self.widget.render(Category.objects),
-                "%s,%s" % (self.cat1.pk, self.cat2.pk))
+                         "%s,%s" % (self.cat1.pk, self.cat2.pk))
         self.assertEqual(self.widget_name.render(Category.objects),
-                u"%s,%s" % (self.cat1.name, self.cat2.name))
-
+                         u"%s,%s" % (self.cat1.name, self.cat2.name))